refactor: update tvFocusable modifier to use FocusState parameter across UI components
