+++ app/src/main/java/com/rdwatch/androidtv/ui/details/components/HeroSection.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.unit.dp
 import com.rdwatch.androidtv.ui.components.SmartTVImageLoader
 import com.rdwatch.androidtv.ui.components.ImagePriority
 import com.rdwatch.androidtv.ui.details.models.*
-import com.rdwatch.androidtv.presentation.components.tvFocusable
 import com.rdwatch.androidtv.ui.focus.TVFocusIndicator
+import com.rdwatch.androidtv.ui.focus.tvFocusable
 
 /**
  * Hero section component for content detail screens
@@ -136,7 +136,7 @@ private fun HeroBackButton(
             modifier = modifier
                 .focusRequester(firstFocusRequester)
                 .tvFocusable(
-                    onFocusChanged = { backButtonFocused = it }
+                    onFocusChanged = { backButtonFocused = it.isFocused }
                 )
         ) {
             Icon(
@@ -490,7 +490,7 @@ private fun HeroActionButton(
             onClick = onClick,
             modifier = Modifier
                 .tvFocusable(
-                    onFocusChanged = { isFocused = it }
+                    onFocusChanged = { isFocused = it.isFocused }
                 )
                 .height(52.dp),
             colors = ButtonDefaults.buttonColors(
